+++ lib/src/countly_flutter.dart
@@ -907,7 +907,7 @@ class Countly {
     return result;
   }
 
-  /// Set to 'true' if you want HTTP POST to be used for all requests
+  /// Set to 'false' if you want HTTP POST not to be used for all requests
   /// Should be call before Countly init
   /// returns the error or success message
   @Deprecated('This functions is deprecated, please use "setHttpPostForced" of CountlyConfig instead')
@@ -1681,7 +1681,7 @@ class Countly {
     return result;
   }
 
-  /// Attempt to send all stored requests to the server.  
+  /// Attempt to send all stored requests to the server.
   Future<void> attemptToSendStoredRequests() async {
     if (!_instance._countlyState.isInitialized) {
       String message = '"initWithConfig" must be called before "attemptToSendStoredRequests"';
@@ -1690,6 +1690,25 @@ class Countly {
     }
     log('Calling "attemptToSendStoredRequests"');
     await _channel.invokeMethod('attemptToSendStoredRequests');
+  }
+
+  /// Add custom headers to all network requests made by the SDK.
+  Future<void> addCustomNetworkRequestHeaders(Map<String, String> customHeaderValues) async {
+    if (!_instance._countlyState.isInitialized) {
+      log('addCustomNetworkRequestHeaders, "initWithConfig" must be called before "addCustomNetworkRequestHeaders"', logLevel: LogLevel.ERROR);
+      return;
+    }
+
+    if (customHeaderValues.isEmpty) {
+      log('addCustomNetworkRequestHeaders, customHeaderValues cannot be empty', logLevel: LogLevel.WARNING);
+      return;
+    }
+
+    log('Calling "addCustomNetworkRequestHeaders" with headers count: [${customHeaderValues.length}]');
+    List<dynamic> args = [];
+    args.add(customHeaderValues);
+
+    await _channel.invokeMethod('addCustomNetworkRequestHeaders', <String, dynamic>{'data': json.encode(args)});
   }
 
   /// starts a timed event